engine: support for image texture

+++ src/image.h
@@ -1,11 +1,12 @@
 #pragma once
 
+#include <stddef.h>
+#include <stdint.h>
 #include "color.h"
-#include "stddef.h"
 typedef struct image
 {
-    size_t width;
-    size_t height;
+    int width;
+    int height;
     Color *converted_data; /* converted data used by the engine */
     void *source_data;     /* raw 8bit data (used by the implementation) */
 } Image;
